swagger

+++ tesodev-korpes/OrderService/internal/types/model.go
@@ -12,7 +12,7 @@ type Order struct {
 	ShipmentStatus   string           `bson:"shipment_status" json:"shipment_status"`
 	PaymentMethod    string           `bson:"payment_method" json:"payment_method"`
 	OrderTotal       int              `bson:"order_total" json:"order_total"`
-	PriceCent        int64            `bson:"priceCent" json:"priceCent"`
+	PriceCent        int              `bson:"priceCent" json:"priceCent"`
 	CreatedAt        time.Time        `bson:"created_at" json:"created_at"`
 	UpdatedAt        time.Time        `bson:"updated_at" json:"updated_at"`
 }
