+++ tesodev-korpes/OrderService/internal/types/type.go
@@ -8,7 +8,7 @@ type OrderRequestModel struct {
 	CustomerId    string `bson:"customer_id" json:"customer_id"`
 	OrderTotal    int    `bson:"order_total" json:"order_total"`
 	PaymentMethod string `bson:"payment_method" json:"payment_method"`
-	PriceCent     int64  `bson:"priceCent" json:"priceCent"`
+	PriceCent     int    `bson:"priceCent" json:"priceCent"`
 	OrderName     string `bson:"order_name" json:"order_name"`
 }
 
@@ -18,7 +18,7 @@ type OrderResponseModel struct {
 	ShipmentStatus string `bson:"shipment_status" json:"shipment_status"`
 	PaymentMethod  string `bson:"payment_method" json:"payment_method"`
 	OrderTotal     int    `bson:"order_total" json:"order_total"`
-	PriceCent      int64  `bson:"priceCent" json:"priceCent"`
+	PriceCent      int    `bson:"priceCent" json:"priceCent"`
 
 	CreatedAt time.Time `bson:"created_at" json:"created_at"`
 	UpdatedAt time.Time `bson:"updated_at" json:"updated_at"`
